tests: Move tests that catch data corruption due to concurrency bugs to racetest.py
We are going to add more such tests and related infrastructure. As the
preparatory step take tests added in

- 5b4dd5f7 (tests: Add test for open vs invalidation race), and
- e923c9a8 (tests: Add test for load vs external invalidation race)

and move them to dedicated module.

No functional changes, only plain code movement.
